Permit specification of alternate CSS template
Right now, I've got an old project with a simulated browser testing environment that does not include URL or Blob. Previously, I was building one of my components using `rollup-plugin-embed-css`, but had to move to a different solution since it was crashing the build.

This patch permits consumers to provide their own `generateCSS` option, for special cases such as this one.